Separate sys.COLORS and sys.fpath into their own files
The sys package forks processes when the package is required to
determine the OS and prefix. Unfortunately, due to a bug in OpenBLAS,
simulatenous calls to fork can deadlock. The torch7 and image packages
require sys at intialization, which now makes deadlocks much more common
when using Torch threads and OpenBLAS.

This moves sys.COLORS and sys.fpath into their own files which can be
required independently, so that torch7 and image don't have to require
all of sys and fork at initialization.